overseer/requirements.txt
flask>=2.0.0
requests>=2.25.0
mysql-connector-python>=8.0.0
